requirements.txt
streamlit
openai
PyMuPDF
python-docx
faiss-cpu
sentence-transformers
python-dotenv
modal==0.62.209
llama-index
streamlit-pills
psutil
streamlit-pdf-viewer